+++ src/main/java/binbean/binbean_BE/controller/CCTVProcessingController.java
@@ -2,7 +2,7 @@ package binbean.binbean_BE.controller;
 
 import binbean.binbean_BE.dto.FloorList;
 import binbean.binbean_BE.dto.response.FloorPlanResponse;
-import binbean.binbean_BE.service.SeatDetectionService;
+import binbean.binbean_BE.service.SeatMappingService;
 import java.io.IOException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -17,16 +17,16 @@ import org.springframework.web.multipart.MultipartFile;
 @RequestMapping("/api/cctv")
 public class CCTVProcessingController {
 
-    private final SeatDetectionService seatDetectionService;
+    private final SeatMappingService seatMappingService;
 
-    public CCTVProcessingController(SeatDetectionService seatDetectionService) {
-        this.seatDetectionService = seatDetectionService;
+    public CCTVProcessingController(SeatMappingService seatMappingService) {
+        this.seatMappingService = seatMappingService;
     }
 
     @PostMapping(value = "/detect", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
     public ResponseEntity<FloorPlanResponse> getDetectedItems(@RequestPart(value = "image") MultipartFile image,
         @RequestPart("floorList") FloorList floorList, @RequestPart("floorNumber") int floorNumber) throws IOException {
-        var response = seatDetectionService.getCurrentOccupiedSeats(image, floorList, floorNumber);
+        var response = seatMappingService.getCurrentOccupiedSeats(image, floorList, floorNumber);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 }
